python-scripte/requirements.txt
# requirements.txt
# Produktionsabhängigkeiten
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
defusedxml==0.7.1
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pillow==11.0.0
PyMuPDF==1.24.14
PyPDF2==3.0.1
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.16.0
tqdm==4.67.1
urllib3==2.2.3
whisper==1.1.10
youtube-transcript-api==0.6.3
